backend/requirements.txt
uvicorn==0.16.0
fastapi==0.71.0
fastlab==0.2.0
pymongo==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
requests==2.27.1
elasticsearch==7.16.2